invalid Visitable for acyclic Adapters realized
